requirements/product.txt
# Flask
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0
Flask-APScheduler==1.11.0
flasgger==0.9.5
python-dotenv==0.15.0

# Deployment
gunicorn==20.0.4

# Crawler
openpyxl==3.0.6
requests==2.25.1
xlwt==1.3.0
xlrd==2.0.1
xlutils==2.0.0
bs4==0.0.1
beautifulsoup4==4.9.3

# TEMP
selenium==3.141.0
pandas==1.2.3